// sourcecode/BTrak.Packages/snova-admin-module/admin-module-components/admin-components/src/lib/admin-module/models/hr-models/time-zone.ts
export class TimeZoneModel {
    timeZoneId: string;
    timeZoneName: string;
    timeZoneOffset: string;
    timeZone: string;
    timeZoneAbbreviation: string;
    countryCode: string;
    countryName: string;
    isArchived: boolean;
    timeStamp: any;
    totalCount: number;
    searchText: string;
}